Refector InputView : merge inputview and outputview
